refactor: split fixtures/generator.py by domain (CHAOS-1542)
Split the 3494-line src/dev_health_ops/fixtures/generator.py into per-domain
mixin modules under src/dev_health_ops/fixtures/generators/. The composer in
generator.py inherits from the new mixins so the public import path
(from dev_health_ops.fixtures.generator import SyntheticDataGenerator) and
every existing call site keep working unchanged.

Pure refactor — zero behavior change. All 76 'fixtures' tests pass.
